Update deps and resolve staticcheck messages
This is a general update of the dependencies and also updates go to 1.21

As part of this the Makefile is updated to remove the long deprecated
lint command and instead use staticcheck. Additionally staticcheck has
been run and all messages resolved.

This update does not change any behaviour in bluemonday, all tests remain
the same and continue to pass, and there is no security issue resolved by
this update, it's purely a maintenance update.